+++ src/Store.API/Controllers/ProductController.cs
@@ -23,63 +23,44 @@ public class ProductController : ControllerBase
     [Route("all")]
     public async Task<ActionResult<List<Product>>> GetProductsAsync()
     {
-        try
+        var query = await _productService.GetAllProductsAsync();
+        if (query == null)
         {
-            var query = await _productService.GetAllProductsAsync();
-            if (query == null)
-            {
-                return NotFound();
-            }
+            return NotFound();
+        }
 
-            return Ok(query);
-        }
-        catch (Exception e)
-        {
-            return BadRequest(new ProblemDetails() 
-                { Title = $"Problem getting products: {e.Message}" });
-        }
+        return Ok(query);
     }
 
     [HttpGet]
     [Route("available")]
     public async Task<ActionResult<List<Product>>> GetAvailableProductsAsync()
     {
-        try
-        {
-            var products = await _productService.GetAvailableProductsAsync();
+        var products = await _productService.GetAvailableProductsAsync();
 
-            return Ok(products.ToArray());
-        }
-        catch (Exception e)
-        {
-            return BadRequest(new ProblemDetails() 
-                { Title = $"Problem getting available products {e.Message}" });
-        }
+        return Ok(products);
     }
 
     [HttpGet("{id:int}")]
     public async Task<ActionResult<Product>> GetProductAsync(int id)
     {
-        try
+        if (id < 1)
         {
-            var product = await _productService.GetProductByIdAsync(id);
-            if (product == null)
-            {
-                return BadRequest(new ProblemDetails()
-                    { Title = $"No product {id}" });
-            }
+            return BadRequest(new ProblemDetails() { Title = "Invalid product id." });
+        }
+        
+        var product = await _productService.GetProductByIdAsync(id);
+        if (product == null)
+        {
+            return BadRequest(new ProblemDetails()
+                { Title = $"No product {id}" });
+        }
 
-            return Ok(product);
-        }
-        catch (Exception e)
-        {
-            return BadRequest(new ProblemDetails() 
-                { Title = $"Problem getting a product {id}: {e.Message}" });
-        }
+        return Ok(product);
     }
 
     [HttpPost]
-    public async Task<ActionResult<Product>> CreateProductAsync([FromBody] CreateProductDto productDto)
+    public async Task<ActionResult<Product>> CreateProductAsync(CreateProductDto productDto)
     {
         if (productDto == null)
         {
@@ -87,21 +68,14 @@ public class ProductController : ControllerBase
         }
 
         var product = _mapper.Map<Product>(productDto);
-        try
-        {
-            int id = await _productService.AddProductAsync(product);
-            product.Id = id;
+        int id = await _productService.AddProductAsync(product);
+        product.Id = id;
 
-            return CreatedAtAction(nameof(CreateProductAsync), new { product.Id }, product);
-        }
-        catch (Exception e)
-        {
-            return BadRequest(new ProblemDetails() { Title = $"Problem updating a product: {e.Message}" });
-        }
+        return CreatedAtAction(nameof(CreateProductAsync), new { product.Id }, product);
     }
 
     [HttpPut]
-    public async Task<ActionResult<Product>> UpdateProductAsync([FromBody] UpdateProductDto productDto)
+    public async Task<ActionResult<Product>> UpdateProductAsync(UpdateProductDto productDto)
     {
         if (productDto == null)
         {
@@ -115,30 +89,21 @@ public class ProductController : ControllerBase
         }
 
         _mapper.Map(productDto, product);
-        try
-        {
-            await _productService.UpdateProductAsync(product);
-            return Ok(product);
-        }
-        catch (Exception e)
-        {
-            return BadRequest(new ProblemDetails() 
-                { Title = $"Problem  updating a product {product.Id}: {e.Message}" });
-        }
+        await _productService.UpdateProductAsync(product);
+
+        return Ok(product);
     }
 
     [HttpDelete("{id:int}")]
     public async Task<ActionResult> DeleteProductAsync(int id)
     {
-        try
+        if (id < 1)
         {
-            await _productService.DeleteProductAsync(id);
-            return Ok();
+            return BadRequest(new ProblemDetails() {Title = "Invalid product id."});
         }
-        catch (Exception e)
-        {
-            return BadRequest(new ProblemDetails() 
-                { Title = $"Problem deleting a product {id}: {e.Message}" });
-        }
+        
+        await _productService.DeleteProductAsync(id);
+
+        return Ok();
     }
 }
